requirements.txt
CacheControl==0.12.6
certifi>=2023.7.22
chardet==3.0.4
discord-webhook==0.13.0
et-xmlfile==1.0.1
idna>=3.4
jdcal==1.3
msgpack>=1.0.5
numpy>=1.26.0
oauthlib>=3.2.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.9
pymssql>=2.2.8
PyMySQL>=1.1.0
python-dateutil>=2.8.2
pytz>=2023.3
PyYAML>=6.0.1
redis>=5.0.0
requests>=2.31.0
requests-cache>=1.1.0
requests-futures>=1.0.1
requests-oauthlib>=1.3.1
six>=1.16.0
SQLAlchemy>=2.0.0
tqdm>=4.66.0
urllib3>=2.0.0
